Add files via upload
This is the code for the project me and my team did for this coursework The project was based on handwriting recognition and evaluating performance of different models. The data was munged primarily using PCA to reduce the dimensions of the data, since the data was pixel values. The 2 models we chose to focus upon were Random Forest and Multi-layer Perceptron. MLP performed better, as was expected. I tried playing around with the hyper-parameters of random forest. The challenges that were faced primarily came from reading the data. The earlier dataset was humongous (8 million images of 256*256 pixels) and reading it on the Apache Spark hardware we were provided with was tough (the system crashed). So, we switched the dataset to one with images of lower resolution.